requirements.txt
fastapi[standard]~=0.115.12
sqlmodel~=0.0.24
pydantic_settings~=2.8.1
httpx~=0.28.1
pandas~=2.2.3
alembic~=1.15.2
typer~=0.15.2
black
uvicorn
pydantic-settings~=2.8.1
sqlalchemy~=2.0.40
pytest~=8.3.5
hypercorn
python-dotenv~=1.1.0
psycopg[binary]
psycopg2-binary
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.9